requirements.txt
streamlit
PyGithub
pandas
plotly
python-dotenv

